// src/api/generated/models/projectRead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SoftTrack
 * An open-source, self-hostable issue tracker inspired by Linear.
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectRead {
  id: number;
  team_id: number;
  name: string;
  description?: string | null;
  color: string;
  created_at: string;
}